Address Claude review on PR #1669 and align test with new error message
- Move join_block_text_parts and synthesize_relationship_block_text from
  tasks/embeddings_task.py to utils/embeddings.py so the GraphQL-facing
  vector stores no longer depend on a Celery task module (SRP / utility
  placement).
- Drop the _prefetched_objects_cache introspection in
  synthesize_relationship_block_text and always go through
  ``order_by("id").values_list``; the cache shape is a Django internal
  that has shifted across versions, and the two extra value queries are
  negligible for a Celery task path.
- Fix the misleading docstring in _attach_block_context_sync — there is
  no ``Count(target_annotations)`` annotation in the queryset; the
  descendant count is computed in Python from the prefetched ID set
  (as the inline comment explains).
- Make _apply_dual_embedding_strategy generic in the embeddable type
  (TypeVar bound to HasEmbeddingMixin) and drop the cast that was
  forcing _embed_relationship's Relationship-typed signature through
  the HasEmbeddingMixin-shaped parameter.
- In useJumpToRelationship, track scroll application separately from
  selection so the deep-link retries scrolling once the virtualised
  renderer mounts the source annotation's page. Selection is still
  applied exactly once per relId change, so user-driven edits are not
  fought.
- Update test_embedding_manager assertion to match the new error
  message that now lists relationship_id among the valid parent IDs.
- Replace the ``as any`` in useJumpToRelationship.test.tsx with a
  properly-typed AnnotationLabelType literal to keep the any-baseline
  gate green.
