+++ opencontractserver/tasks/embeddings_task.py
@@ -1,6 +1,6 @@
 import logging
 from concurrent.futures import ThreadPoolExecutor, as_completed
-from typing import Any, Callable, Optional, Union, cast
+from typing import Any, Callable, Optional, TypeVar, Union, cast
 
 import requests
 from celery import shared_task
@@ -8,9 +8,6 @@ from celery.utils.log import get_task_logger
 from django.contrib.auth import get_user_model
 
 from opencontractserver.annotations.models import Annotation, Note, Relationship
-from opencontractserver.constants.annotations import (
-    SUBTREE_GROUP_BLOCK_TEXT_MAX_CHARS,
-)
 from opencontractserver.constants.document_processing import EMBEDDING_API_BATCH_SIZE
 from opencontractserver.corpuses.models import Corpus
 from opencontractserver.documents.models import Document
@@ -26,6 +23,9 @@ from opencontractserver.pipeline.utils import (
 )
 from opencontractserver.shared.mixins import HasEmbeddingMixin
 from opencontractserver.types.enums import ContentModality
+from opencontractserver.utils.embeddings import (
+    synthesize_relationship_block_text,
+)
 
 User = get_user_model()
 
@@ -185,13 +185,16 @@ class EmbeddingGenerationError(Exception):
     pass
 
 
+_EmbeddableT = TypeVar("_EmbeddableT", bound=HasEmbeddingMixin)
+
+
 def _apply_dual_embedding_strategy(
-    obj: HasEmbeddingMixin,
+    obj: _EmbeddableT,
     text: str,
     corpus_id: Optional[int],
     obj_type: str,
     obj_id: int,
-    embed_func: Callable[[HasEmbeddingMixin, BaseEmbedder, str], bool],
+    embed_func: Callable[[_EmbeddableT, BaseEmbedder, str], bool],
 ) -> None:
     """
     Apply the dual embedding strategy to any embeddable object.
@@ -951,91 +954,6 @@ def calculate_embedding_for_note_text(
 # --------------------------------------------------------------------------- #
 
 
-def join_block_text_parts(
-    chunks: list[str],
-    *,
-    max_chars: int = SUBTREE_GROUP_BLOCK_TEXT_MAX_CHARS,
-) -> str:
-    """Newline-join non-empty strings, truncated at ``max_chars``.
-
-    Shared by every code path that surfaces an OC_SUBTREE_GROUP's block
-    text — the embedder (so the vector reflects what we'll later show),
-    the relationship-vector-store result shaper, and the annotation
-    vector store's block-context attach. Centralising here keeps the
-    cap/truncation rules from drifting across three call sites.
-
-    Empty strings are skipped so a partially-parsed structural annotation
-    with no raw text yet doesn't inject stray newlines.
-    """
-    parts: list[str] = []
-    running = 0
-    for chunk in chunks:
-        if not chunk:
-            continue
-        if running == 0:
-            if len(chunk) >= max_chars:
-                parts.append(chunk[:max_chars])
-                running = max_chars
-                break
-            parts.append(chunk)
-            running = len(chunk)
-            continue
-        # ``+1`` accounts for the newline separator the join below will
-        # insert between successive non-empty parts.
-        if running + 1 + len(chunk) > max_chars:
-            remaining = max_chars - running - 1
-            if remaining > 0:
-                parts.append(chunk[:remaining])
-            break
-        parts.append(chunk)
-        running += 1 + len(chunk)
-
-    return "\n".join(parts)
-
-
-def synthesize_relationship_block_text(
-    relationship: Relationship,
-    *,
-    max_chars: int = SUBTREE_GROUP_BLOCK_TEXT_MAX_CHARS,
-) -> str:
-    """Build the embedder-facing string for a ``Relationship``.
-
-    The string is ``source.raw_text`` followed by each target's ``raw_text``
-    on its own line, truncated to ``max_chars``. Target ordering is by ID so
-    re-embedding the same relationship produces a stable input string — that
-    way ``add_embedding``'s idempotent upsert short-circuits unchanged
-    inputs instead of overwriting equivalent vectors.
-
-    Delegates the cap/truncation logic to :func:`join_block_text_parts`, which
-    is also called from the GraphQL-facing vector-store paths so all three
-    code paths produce byte-identical output for the same inputs.
-
-    When the caller has prefetched ``source_annotations`` /
-    ``target_annotations`` with ``Prefetch(queryset=Annotation.objects.order_by("id"))``,
-    we read ``raw_text`` from the prefetched objects directly so the
-    helper costs zero extra queries; otherwise the ``order_by("id")``
-    falls back to fresh DB hits with deterministic ordering.
-    """
-    src_qs = relationship.source_annotations
-    tgt_qs = relationship.target_annotations
-    prefetched = getattr(relationship, "_prefetched_objects_cache", None) or {}
-    if "source_annotations" in prefetched:
-        sources: list[str] = [a.raw_text or "" for a in src_qs.all()]
-    else:
-        sources = [
-            (text or "")
-            for text in src_qs.order_by("id").values_list("raw_text", flat=True)
-        ]
-    if "target_annotations" in prefetched:
-        targets: list[str] = [a.raw_text or "" for a in tgt_qs.all()]
-    else:
-        targets = [
-            (text or "")
-            for text in tgt_qs.order_by("id").values_list("raw_text", flat=True)
-        ]
-    return join_block_text_parts([*sources, *targets], max_chars=max_chars)
-
-
 def _embed_relationship(
     relationship: Relationship,
     embedder: BaseEmbedder,
@@ -1212,10 +1130,7 @@ def calculate_embeddings_for_relationship_batch(
                 corpus_id=int(corpus_id) if corpus_id else None,
                 obj_type="relationship",
                 obj_id=rel.id,
-                embed_func=cast(
-                    "Callable[[HasEmbeddingMixin, BaseEmbedder, str], bool]",
-                    _embed_relationship,
-                ),
+                embed_func=_embed_relationship,
             )
             result["succeeded"] += 1
         except Exception as e:
