+++ opencontractserver/utils/embeddings.py
@@ -1,13 +1,105 @@
 import logging
-from typing import Optional, Union, cast
+from typing import TYPE_CHECKING, Optional, Union, cast
 
 from channels.db import database_sync_to_async
 
+from opencontractserver.constants.annotations import (
+    SUBTREE_GROUP_BLOCK_TEXT_MAX_CHARS,
+)
 from opencontractserver.pipeline.base.embedder import BaseEmbedder
 from opencontractserver.pipeline.base.file_types import FileTypeEnum
 
+if TYPE_CHECKING:
+    from opencontractserver.annotations.models import Relationship
+
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.DEBUG)
+
+
+def join_block_text_parts(
+    chunks: list[str],
+    *,
+    max_chars: int = SUBTREE_GROUP_BLOCK_TEXT_MAX_CHARS,
+) -> str:
+    """Newline-join non-empty strings, truncated at ``max_chars``.
+
+    Shared by every code path that surfaces an OC_SUBTREE_GROUP's block
+    text — the embedder (so the vector reflects what we'll later show),
+    the relationship-vector-store result shaper, and the annotation
+    vector store's block-context attach. Centralising here keeps the
+    cap/truncation rules from drifting across three call sites.
+
+    Empty strings are skipped so a partially-parsed structural annotation
+    with no raw text yet doesn't inject stray newlines.
+    """
+    parts: list[str] = []
+    running = 0
+    for chunk in chunks:
+        if not chunk:
+            continue
+        if running == 0:
+            if len(chunk) >= max_chars:
+                parts.append(chunk[:max_chars])
+                running = max_chars
+                break
+            parts.append(chunk)
+            running = len(chunk)
+            continue
+        # ``+1`` accounts for the newline separator the join below will
+        # insert between successive non-empty parts.
+        if running + 1 + len(chunk) > max_chars:
+            remaining = max_chars - running - 1
+            if remaining > 0:
+                parts.append(chunk[:remaining])
+            break
+        parts.append(chunk)
+        running += 1 + len(chunk)
+
+    return "\n".join(parts)
+
+
+def synthesize_relationship_block_text(
+    relationship: "Relationship",
+    *,
+    max_chars: int = SUBTREE_GROUP_BLOCK_TEXT_MAX_CHARS,
+) -> str:
+    """Build the embedder-facing string for a ``Relationship``.
+
+    The string is ``source.raw_text`` followed by each target's ``raw_text``
+    on its own line, truncated to ``max_chars``. Target ordering is by ID so
+    re-embedding the same relationship produces a stable input string — that
+    way ``add_embedding``'s idempotent upsert short-circuits unchanged
+    inputs instead of overwriting equivalent vectors.
+
+    Delegates the cap/truncation logic to :func:`join_block_text_parts`, which
+    is also called from the GraphQL-facing vector-store paths so all three
+    code paths produce byte-identical output for the same inputs.
+
+    When the caller has prefetched ``source_annotations`` /
+    ``target_annotations`` with ``Prefetch(queryset=Annotation.objects.order_by("id"))``,
+    we read ``raw_text`` from the prefetched objects directly so the
+    helper costs zero extra queries; otherwise the ``order_by("id")``
+    falls back to fresh DB hits with deterministic ordering.
+    """
+    # Always hit the DB via ``order_by("id").values_list`` rather than
+    # introspecting Django's private ``_prefetched_objects_cache``. Subtree
+    # groups are small relative to leaf annotations and this is a Celery
+    # task path, so two extra value queries per relationship are
+    # negligible; in exchange we avoid coupling to a Django internal whose
+    # shape has shifted across versions.
+    sources = [
+        (text or "")
+        for text in relationship.source_annotations.order_by("id").values_list(
+            "raw_text", flat=True
+        )
+    ]
+    targets = [
+        (text or "")
+        for text in relationship.target_annotations.order_by("id").values_list(
+            "raw_text", flat=True
+        )
+    ]
+    return join_block_text_parts([*sources, *targets], max_chars=max_chars)
 
 
 def get_embedder(
